+++ src/main/java/com/monumental/repositories/MonumentRepository.java
@@ -1,4 +1,4 @@
-package com.monumental.services;
+package com.monumental.repositories;
 
 import com.monumental.models.Monument;
 import org.springframework.data.repository.CrudRepository;
